refactor: extract BattleContext types from conflict detection (#4284)
Slice A of combat phase 4: move BattleContext and AttackingSide into
battle_context.dart; re-export from barrel and conflict_detection for
stable imports. detectConflicts stays detection-only.

Refs #4284
